keep only three pages in index page.
